feat: add preloader

+++ src/views/TaskViewPage.vue
@@ -1,23 +1,30 @@
 <template lang="pug">
-  .container.task-view-page
-    h1.header Задача \#{{ task.id }}
-    span.header__sub
-    .task-view-page__content
-      small Заголовок:
-      p {{ task.title }}
-      small Описание:
-      p {{ task.description }}
-      small Дата создания:
-      p {{ task.created_date }}
-      select.form-control__select(v-model="task.status" @change="handleChangeStatus")
-        option(v-for="option in statuses" :key="option.id" :value="option.id") {{ option.title }}
-    button.btn.btn--danger(@click="deleteTask" style="margin-top: 20px") Удалить
+  .container(v-loading="isLoading")
+    preloader(v-if="isLoading")
+    .task-view-page(v-else="isLoading")
+      h1.header Задача \#{{ task.id }}
+      span.header__sub
+      .task-view-page__content
+        small Заголовок:
+        p {{ task.title }}
+        small Описание:
+        p {{ task.description }}
+        small Дата создания:
+        p {{ task.created_date }}
+        small Статус:
+        p
+          select.form-control__select(v-model="task.status" @change="handleChangeStatus")
+            option(v-for="option in statuses" :key="option.id" :value="option.id") {{ option.title }}
+      button.btn.btn--danger(@click="deleteTask" style="margin-top: 20px") Удалить
 </template>
 
 <script>
+import Preloader from '@/components/Preloader';
 import statuses from '@/utils/statuses';
+
 export default {
     name: 'TaskViewPage',
+    components: { Preloader },
     data() {
         return {
             isLoading: false,
@@ -40,10 +47,15 @@ export default {
             } catch (e) {
                 console.error(e);
             } finally {
-                this.isLoading = false;
+                setTimeout(() => {
+                    this.isLoading = false;
+                }, 1500);
             }
         },
         async deleteTask() {
+            if (!confirm('Точно удалить?')) {
+                return;
+            }
             this.isLoading = true;
             try {
                 const {status} = await this.axios.delete(`/tasks/${this.task.id}`);
